src/tasks/python/preprocessing/requirements.txt
cloud-volume
numpy
scipy
tqdm
shapely
Pillow
google-cloud-pubsub